port to typescript and template socketIO interaction
