+++ package.json
@@ -5,23 +5,26 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js",
+    "start": "npx tsc && node dist/index.js",
     "dev": "nodemon index"
   },
   "author": "Adam Kagel, Maclean Hartford",
   "license": "ISC",
-  "type": "module",
   "dependencies": {
-    "@types/express": "^4.17.9",
     "dotenv": "^8.2.0",
     "ejs": "^3.1.5",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
-    "multer": "^1.4.2"
+    "multer": "^1.4.2",
+    "socket.io": "^3.1.0"
   },
   "devDependencies": {
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.9",
     "@types/express-session": "^1.17.3",
     "@types/multer": "^1.4.5",
-    "nodemon": "^2.0.6"
+    "@types/socket.io": "^2.1.12",
+    "nodemon": "^2.0.6",
+    "typescript": "^4.1.3"
   }
 }
