+++ tsconfig.json
@@ -1,0 +1,18 @@
+{
+	"compilerOptions": {
+		"target": "es6",
+        "module": "commonjs",
+		"sourceMap": true,
+		"outDir": "dist",
+		"baseUrl": "./",
+		"esModuleInterop": true,
+		"paths": {
+			"*": [
+				"node_modules/*"
+			]
+		}
+	},
+	"include": [
+		"src/**/*"
+	]
+}
